Make function instead of fixture
